Remove UDB bootloader version memory section for generic HIC
The UDB IF needs to know the UDB BL version. This version is different
from the DAPLINK_VERSION, it is a string that contains info about the
last git commit, local modifications and can be injected through a CI
pipeline.

However, this doesn't need to be in the generic STM32H743xx HIC and this
frees up additional flash.